+++ src/main/java/br/com/gerenciadorcontatos/repository/EnderecoRepository.java
@@ -1,11 +1,11 @@
 package br.com.gerenciadorcontatos.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import br.com.gerenciadorcontatos.model.entity.Endereco;
 
 @Repository
-public interface EnderecoRepository extends JpaRepository<Endereco, Long>{
-
-}
+public interface EnderecoRepository extends JpaRepository<Endereco, BigInteger>{}
