+++ src/main/java/br/com/gerenciadorcontatos/service/ContatoService.java
@@ -1,16 +1,17 @@
 package br.com.gerenciadorcontatos.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface ContatoService<T> {
 	
 	T create(T obj);
 
-	T findById(Long id);
+	T findById(BigInteger id);
 
 	List<T> findAll();
 
 	boolean update(T obj);
 	
-	boolean delete(Long id);
+	boolean delete(BigInteger id);
 }
